Tue, Mar 14, 2023, 2:28 PM +01:00

+++ movies/11.sql
@@ -1,7 +1,7 @@
 SELECT title FROM movies WHERE id IN
 (SELECT movie_id FROM ratings
 WHERE movie_id IN (SELECT movie_id FROM stars
-WHERE person_id IN (SELECT id FROM people WHERE name = "Chadwick Boseman")) ORDER BY CONVERT(FLOAT,rating) DESC LIMIT 5);
+WHERE person_id IN (SELECT id FROM people WHERE name = "Chadwick Boseman")) ORDER BY CONVERT(int,rating) DESC LIMIT 5);
 
 
 SELECT movie_id FROM ratings WHERE movie_id IN
